Use proper file name for v6 dict and tidy up tests
