+++ RapidOcrNet/RapidOcrNet.csproj
@@ -71,13 +71,10 @@
 		<None Update="models\v6\PP-OCRv6_rec_medium.onnx">
 			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
+		<None Update="models\v6\ppocrv6_dict.txt">
+		  <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
 		<None Update="models\v6\ppocrv6_tiny_dict.txt">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
-		<None Update="models\v6\ppocrv6_small_dict.txt">
-			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-		</None>
-		<None Update="models\v6\ppocrv6_medium_dict.txt">
 			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
 	</ItemGroup>
